moved config code to not import torch
