+++ optimum_benchmark/backends/neural_compressor/config.py
@@ -4,7 +4,7 @@ from typing import Any, Dict
 
 from omegaconf import OmegaConf
 
-from ..base import BackendConfig
+from ..config import BackendConfig
 
 OmegaConf.register_new_resolver(
     "neural_compressor_version",
