+++ optimum_benchmark/backends/pytorch/config.py
@@ -1,11 +1,11 @@
-import importlib.metadata
 import os
 from dataclasses import dataclass, field
 from typing import Any, Dict, Optional
 
 from omegaconf import OmegaConf
 
-from ..base import BackendConfig
+from ...import_utils import torch_version
+from ..config import BackendConfig
 
 OmegaConf.register_new_resolver(
     "device_count",
@@ -15,10 +15,7 @@ OmegaConf.register_new_resolver(
     "is_inference",
     lambda benchmark_name: benchmark_name == "inference",
 )
-OmegaConf.register_new_resolver(
-    "pytorch_version",
-    lambda: importlib.metadata.version("torch"),
-)
+OmegaConf.register_new_resolver("pytorch_version", lambda: torch_version())
 
 DEVICE_MAPS = ["auto", "sequential"]
 AMP_DTYPES = ["bfloat16", "float16"]
