+++ optimum_benchmark/experiment.py
@@ -2,38 +2,38 @@ import os
 import platform
 from dataclasses import dataclass, field
 from logging import getLogger
-from typing import Any, Dict, Type
+from typing import TYPE_CHECKING, Any, Dict, Type
 
 import hydra
-from accelerate import __version__ as accelerate_version
-from diffusers import __version__ as diffusers_version
 from hydra.core.config_store import ConfigStore
 from hydra.utils import get_class
 from omegaconf import DictConfig, OmegaConf
-from optimum.exporters import TasksManager
-from optimum.version import __version__ as optimum_version
-from transformers import __version__ as transformers_version
 
-from .backends.base import Backend
 from .backends.neural_compressor.config import INCConfig
 from .backends.onnxruntime.config import ORTConfig
 from .backends.openvino.config import OVConfig
 from .backends.pytorch.config import PyTorchConfig
-from .benchmarks.base import Benchmark
-from .benchmarks.inference import InferenceConfig
-from .benchmarks.training import TrainingConfig
+from .benchmarks.inference.config import InferenceConfig
+from .benchmarks.training.config import TrainingConfig
 from .env_utils import get_cpu, get_cpu_ram_mb
+from .import_utils import (
+    accelerate_version,
+    diffusers_version,
+    optimum_version,
+    transformers_version,
+)
+from .task_utils import infer_task
+
+if TYPE_CHECKING:
+    from .backends.base import Backend
+    from .benchmarks.base import Benchmark
 
 LOGGER = getLogger("experiment")
 
 OmegaConf.register_new_resolver(
     "infer_task",
-    # TODO: find a better way for this; it doesn't
-    # always work because it relies on hub metadata
-    lambda model, revision: TasksManager.infer_task_from_model(
-        model=model,
-        revision=revision,
-    ),
+    # TODO: find a better way for this; it doesn't always work because it relies on hub metadata
+    lambda model, revision: infer_task(model, revision),
 )
 
 
@@ -68,10 +68,10 @@ class ExperimentConfig:
     # TODO: add gpu info when available
     environment: Dict = field(
         default_factory=lambda: {
-            "optimum_version": optimum_version,
-            "transformers_version": transformers_version,
-            "accelerate_version": accelerate_version,
-            "diffusers_version": diffusers_version,
+            "optimum_version": optimum_version(),
+            "transformers_version": transformers_version(),
+            "accelerate_version": accelerate_version(),
+            "diffusers_version": diffusers_version(),
             "python_version": platform.python_version(),
             "system": platform.system(),
             "cpu": get_cpu(),
@@ -101,8 +101,8 @@ def run_experiment(experiment: DictConfig) -> None:
     OmegaConf.save(experiment, "hydra_config.yaml", resolve=True)
 
     # Allocate requested backend
-    backend_factory: Type[Backend] = get_class(experiment.backend._target_)
-    backend: Backend = backend_factory(
+    backend_factory: Type["Backend"] = get_class(experiment.backend._target_)
+    backend: "Backend" = backend_factory(
         task=experiment.task,
         model=experiment.model,
         device=experiment.device,
@@ -116,8 +116,8 @@ def run_experiment(experiment: DictConfig) -> None:
         raise e
 
     # Allocate requested benchmark
-    benchmark_factory: Type[Benchmark] = get_class(experiment.benchmark._target_)
-    benchmark: Benchmark = benchmark_factory()
+    benchmark_factory: Type["Benchmark"] = get_class(experiment.benchmark._target_)
+    benchmark: "Benchmark" = benchmark_factory()
     try:
         benchmark.configure(experiment.benchmark)
     except Exception as e:
