+++ optimum_benchmark/import_utils.py
@@ -1,22 +1,51 @@
+import importlib.metadata
 import importlib.util
 
+_transformers_available = importlib.util.find_spec("transformers") is not None
+_accelerate_available = importlib.util.find_spec("accelerate") is not None
+_diffusers_available = importlib.util.find_spec("diffusers") is not None
+_optimum_available = importlib.util.find_spec("optimum") is not None
 _torch_available = importlib.util.find_spec("torch") is not None
 _onnxruntime_available = importlib.util.find_spec("onnxruntime") is not None
-_is_openvino_available = importlib.util.find_spec("openvino") is not None
-_is_neural_compressor_available = importlib.util.find_spec("neural_compressor") is not None
+_openvino_available = importlib.util.find_spec("openvino") is not None
+_neural_compressor_available = importlib.util.find_spec("neural_compressor") is not None
 
 
-def is_torch_available():
-    return _torch_available
+def torch_version():
+    if _torch_available:
+        return importlib.metadata.version("torch")
 
 
-def is_onnxruntime_available():
-    return _onnxruntime_available
+def onnxruntime_version():
+    if _onnxruntime_available:
+        return importlib.metadata.version("onnxruntime")
 
 
-def is_openvino_available():
-    return _is_openvino_available
+def openvino_version():
+    if _openvino_available:
+        return importlib.metadata.version("openvino")
 
 
-def is_neural_compressor_available():
-    return _is_neural_compressor_available
+def neural_compressor_version():
+    if _neural_compressor_available:
+        return importlib.metadata.version("neural_compressor")
+
+
+def optimum_version():
+    if _optimum_available:
+        return importlib.metadata.version("optimum")
+
+
+def transformers_version():
+    if _transformers_available:
+        return importlib.metadata.version("transformers")
+
+
+def accelerate_version():
+    if _accelerate_available:
+        return importlib.metadata.version("accelerate")
+
+
+def diffusers_version():
+    if _diffusers_available:
+        return importlib.metadata.version("diffusers")
